requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
psycopg2-binary==2.9.1
bcrypt==3.2.0
PyJWT==2.3.0
Werkzeug==2.0.1
gunicorn==20.1.0
boto3==1.26.0 